FA-23 - refactor(UI_library): Improve the code and test descriptions
