+++ packages/ui_library/lib/widgets/u_avatars_profile/u_app_avatar_profile/u_app_avatar_profile.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:ui_library/core/core_export.dart';
 import 'package:ui_library/ui_library_export.dart';
 import 'package:ui_library/widgets/global/clipper/u_clipper.dart';
-import 'package:ui_library/widgets/global/placeholders/placeholder.dart';
+import 'package:ui_library/widgets/global/placeholder.dart';
 import 'package:ui_library/widgets/global/unread_messages_indicator.dart';
 
 import '../models/models_export.dart';
