+++ packages/ui_library/lib/widgets/u_user_profile/u_user_profile_notifications/u_user_profile_notification.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:ui_library/widgets/global/clipper/u_clipper.dart';
-import 'package:ui_library/widgets/global/placeholders/placeholder.dart';
+import 'package:ui_library/widgets/global/placeholder.dart';
 
 import '../../../core/core_export.dart';
 import '../../u_text/u_text_export.dart';
