+++ packages/ui_library/test/core/const/u_fonts_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:ui_library/core/const/const_export.dart';
 
 void main() {
-  group('Test UFonts class |', () {
+  group('Unit Test UFonts class |', () {
     test('Should return the correct Icon font', () {
       const _iconsFont = 'UIcons';
       expect(UFonts.iconsFont, _iconsFont);
