+++ packages/ui_library/test/core/const/u_sizes_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:ui_library/core/const/const_export.dart';
 
 void main() {
-  group('Test USizes class |', () {
+  group('Unit Test USizes class |', () {
     test('Should return correct size for UIcons', () {
       const _iconSize = 24.0;
       expect(USizes.iconSize, _iconSize);
