+++ packages/ui_library/test/core/theme/models/u_themes_test.dart
@@ -3,9 +3,11 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:ui_library/core/theme/models/u_themes.dart';
 
 void main() {
-  test('Test UThemes class', () {
-    expect(UThemes.dark, isA<ThemeData>());
-    expect(UThemes.light, isA<ThemeData>());
-    expect(UThemes.tbd, isA<ThemeData>());
+  group('Unit Test UThemes class |', () {
+    test('Should return correct UThemes type', () {
+      expect(UThemes.dark, isA<ThemeData>());
+      expect(UThemes.light, isA<ThemeData>());
+      expect(UThemes.tbd, isA<ThemeData>());
+    });
   });
 }
