+++ packages/ui_library/test/widgets/u_user_profile/models/u_user_profile_sizes_test.dart
@@ -2,15 +2,15 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:ui_library/ui_library_export.dart';
 
 void main() {
-  group('UUserProfileSizes |', () {});
+  group('Unit Test UUserProfileSizes |', () {});
 
-  test('Test the sizes for User Profiles Widgets', () {
+  test('Should return correct sizes for User Profiles Widgets', () {
     const _largeSize = 56;
     const _normalSize = 40;
     const _topMenuBarSize = 30;
 
-    expect(UUserProfileSize.large, _largeSize);
-    expect(UUserProfileSize.normal, _normalSize);
-    expect(UUserProfileSize.topMenuBar, _topMenuBarSize);
+    expect(UUserProfileSize.large.size, _largeSize);
+    expect(UUserProfileSize.normal.size, _normalSize);
+    expect(UUserProfileSize.topMenuBar.size, _topMenuBarSize);
   });
 }
