+++ packages/ui_library/test/widgets/u_user_profile/u_user_profile_status/u_user_profile_status_clipper_golden_widget_test.dart
@@ -1,8 +1,6 @@
 import 'package:alchemist/alchemist.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:ui_library/core/const/const_export.dart';
-import 'package:ui_library/widgets/global/clipper/u_clipper.dart';
 import 'package:ui_library/widgets/u_user_profile/models/models_export.dart';
 import 'package:ui_library/widgets/u_user_profile/u_user_profile_status/u_user_profile_status.dart';
